Use enumeratum IntEnum for generated enums
See https://github.com/47deg/pbdirect/pull/20

Because it is still a plain old sealed abstract class + case object
hierarchy, this change will not affect Avro serialization using avro4s
(confirmed with manual testing).

Note: This change assumes that enumeratum is on Mu's classpath, which is
not yet true. PR on the way to add it as a dependency.