Proof-of-concept using scalameta for codegen
I find this easier to work with than the existing printer, which is
basically doing fancy string concatenation, because scalameta gives us a
nice AST with which to model the code. It also provides quasiquotes,
which are very convenient, and it indents the code correctly for us.

(Unfortunately we end up losing a lot of the type information when we
do a catamorphism, and we have to add it back in using casting. I
haven't found a nice way around this. Open to suggestions.)

If people are happy with this approach, I suggest we remove
`higherkindness.skeuomorph.mu.print` and update the Mu sbt plugin to use
the new scalameta-based codegen. In future we could also migrate the
OpenAPI codegen to scalameta.