Split useIdiomaticEndpoints into useIdiomaticGrpc and useIdiomaticScala
This PR addresses [sbt-mu-srcgen #93](https://github.com/higherkindness/sbt-mu-srcgen/issues/93) along with an upcoming code-change PR for `sbt-mu-srcgen` and a documentation update PR for `mu-scala`.

It does so by splitting the `useIdiomaticEndpoints` source-generation flag, created in [mu-scala #599](https://github.com/higherkindness/mu-scala/issues/599) to control service method capitalization and package prefixes, into 2 distinct flags with different defaults:

1) `useIdiomaticGrpc` which controls the namespace prefix and defaults to `true`;
2) `useIdiomaticScala` which controls the endpoint capitalization and defaults to `false`.

More formally, if `useIdiomaticScala` is true and if _all_ the RPC call definitions in an input IDL are capitalized, the Scala methods in the services generated from this IDL will be _de_-capitalized, while the gRPC endpoints derived from those will be re-capitalized to match the original IDL. For example, a Protobuf IDL `rpc GetBook` definition becomes a Scala `def getBook` method inside a `methodNameStyle = Capitalized` service, which then generates a `GetBook` gRPC endpoint. A `getBook` IDL definition (for example from Avro) would remain unchanged (`methodNameStyle = Unchanged`, `def getBook`, and `getBook` gRPC endpoint).

The reason for the limitation that all RPC calls must be capitalized for the option to apply, is that we do not annotate individual methods, only the top-level `@service` annotation. Thus if we were to de-capitalize a mixed-capitalization group of calls and apply `methodNameStyle = Capitalized` to the service, some of the originally-lowercased methods would be incorrectly re-capitalized which is part of the issue we're addressing here (for example, `getBook` and `GetBook` would both become `def getBook` Scala and `GetBook` endpoints).

Note that the PR also moves some test resources into the `resource` root, and normalizes them a bit between the Avro and Proto cases.

I'm happy to provide some more background on the "idiomatic" flag's history as I understand it, in case it helps in reviewing this change. Feel free to ask about this or anything else unclear.